requirements.txt
# API dependencies
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
typing-extensions==4.8.0

# LLM dependencies
openai==1.3.0

# Frontend dependencies
streamlit==1.41.0
streamlit-chat==0.1.0

# Other dependencies
Flask==2.3.1
MarkupSafe==2.1.2
Flask-Session==0.8.0


